admin: admin-lp-git-permissions -- allow development opening to be overridden
We have a chicken and egg situation.  We cannot mark opening:
repositories: true until we have repositories created by
admin-seed-repositories _and_ we have run admin-lp-git-permissions over
it.  But we also do not want to try and set permissions for development
series until it is marked open.  Add a new --initialise-development flag
to allow the default opening check to be ignored for this one-time
initialisation.

Signed-off-by: Andy Whitcroft <apw@canonical.com>