ci: remove branch namespace check on pr
Checking the branch name at PR time is too late to be useful. We should
instead enforce this at push time. Unfortunately, Actions are not able
to reject pushes and CI fails on push are silent which makes the check
pointless. A followup feature to add a pre-receive hook is being
considered. Until then there is no enforcement on branch names.

Signed-off-by: Cory Todd <cory.todd@canonical.com>