cranky: pull-sources: add wrapper around pull-source
cranky pull-source works with a single package, whereas most other cranky
commands work with a handle. Most of the time, crankers want to download
all the packages for a given kernel.

Also, cranky pull-source will download the latest version it finds in the
archive, but only if packages are in the archive. If they are in a PPA, the
version needs go be given and they are not the same for every package.

Instead of looking for versions in a specific pocket, cranky-pull-sources
look at the changelog of checked out trees and tries to download the
previos version. This covers the use case where crankers will review
packages after preparing and building packages (similarly reviewers will
have checked out and reviewed the prepared tree).

This will fail the case where a tree had to prepared again due to a build
failure or the package to be reviewed against is not the previous version
as in the changelog. Still, this should work for the most common cases.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>