packaging: setup linter framework
Add support for modern Python and linting tools. We use two files to
control the whole lint and formatting system.

.pre-commit-config.yaml, as the name applies, controls a tool that runs
commands prior to commit. This is nice but the real motivation for using
pre-commit is that it reliably routes files to the correct tool by file
type. e.g. don't pass bash scripts to Python linters. Since most of our
scripts are extensionless and most linters decide what to lint based on
file extensions, pre-commit's solution is ideal. Only modified files
will be routed to their respective linters or formatters.

pyproject.toml is an attempt at standardising the _many_ flavours of
python metadata files. For our purposes, it consolidates all dependency
management and lint/format configuration into a single file. The file is
intentionally written to not attempt to install kteam-tools as a
package.

Signed-off-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: Andrei Gherzan <andrei.gherzan@canonical.com>
Acked-by: Andrea Righi <andrea.righi@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>