create-kernel-tasks: add support for additional cycle qualifiers
We currently support stable cycles (2023.05.15) and development cycles
with a "d" prefix (d2023.05.15).  We would like to add separate cycle
qualifiers for the security sub-cycles and (later) for private CRD
cycles.

Switch the "cycle-date" parameter to a "cycle" parameter with any
associated qualifying prefix.  All cycles are expected to be aligned by
default; other than development ("d") which is free form.  We maintain
compatibility with the existing `--devel` form for existing users.

Signed-off-by: Andy Whitcroft <apw@canonical.com>