ktl/kernel_series: support gzip compressed json formatted data
We use YAML form for kernel-series because it is significantly easier
for a human to parse.  It also allows reuse of common data helping to
reduce the size of the data itself.  However it turns out that decoding
large YAML files is horribly expensive taking approximatly a full second
on an Thinkpad x250 laptop.  Converting the same file mechanically to
JSON form creates a file 2.5x the size, but this still decodes 8x
faster.  By compressing this using gzip we more than make up for that
size increase and significantly improve KernelSeries() instantiation
time by close to 5x.

We can maintain the convienience of a simple YAML format file for human
manipulation by maintaining a rapid mirror of the git based YAML file in
a compressed JSON form.  By expressing this data in our new PS5
envionment we also gain further benefits from avoiding git.l.n
permissions checks.  The only negative is introduction of a short (of
the order of a minute) delay to updates to kernel-series.

Detailed analysis.  Taking a snapshot of kernel-series.yaml for
comparison, it is current 285Kb and a raw conversion
(json.dumps(yaml.safe_load())) is 716Kb:

kernel-series.yaml          285585
kernel-series.json          716432

Loading and parsing the kernel-series.yaml takes about 1.17s (averaged
over 10 loads):

1: ./test7.py
	    Mean        Std.Dev.    Min         Median      Max
real        1.174       0.082       1.095       1.143       1.331
user        1.130       0.067       1.066       1.104       1.262
sys         0.029       0.010       0.012       0.028       0.044

loading and parsing the kernel-series.json take about 0.12s (over 10
runs) dispite being 2.5x times the size due to routing-table duplication
in the conversion:

1: ./test7.py
	    Mean        Std.Dev.    Min         Median      Max
real        0.123       0.016       0.111       0.116       0.166
user        0.103       0.007       0.094       0.101       0.118
sys         0.016       0.005       0.008       0.016       0.029

Things are obviously worse when this data is remote in the DC, loading
kernel-series.yaml from the DC, costing an average of .22s to download:

1: ./test7.py
	    Mean        Std.Dev.    Min         Median      Max
real        1.389       0.118       1.307       1.347       1.730
user        1.083       0.047       1.019       1.072       1.183
sys         0.033       0.026       0.012       0.020       0.096

Loading loading kernel-series.json from the DC, costing an average of .36s
to download due to its size:

1: ./test7.py
	    Mean        Std.Dev.    Min         Median      Max
real        0.478       0.018       0.460       0.474       0.522
user        0.116       0.013       0.102       0.113       0.138
sys         0.019       0.009       0.004       0.021       0.031

It should be noted that the kernel-series data is highly repetative and
as such is highly compressible, compressing to 2% of its original size:

kernel-series.json          716432
kernel-series.json.gz        20635

Loading and decompressing kernel-series.json.gz locally is very close to the
same performance as the uncompressed load, and indeed the difference here is
within the noise of repeated runs:

1: ./test7.py
	    Mean        Std.Dev.    Min         Median      Max
real        0.131       0.009       0.121       0.128       0.149
user        0.115       0.009       0.096       0.115       0.125
sys         0.015       0.009       0.000       0.014       0.027

Loading this kernel-series.json.gz from the DC is significantly faster,
reducing the remote overhead to 0.16s:

1: ./test7.py
	    Mean        Std.Dev.    Min         Median      Max
real        0.287       0.015       0.272       0.282       0.327
user        0.129       0.014       0.108       0.127       0.159
sys         0.015       0.006       0.004       0.015       0.025

Finally it should be noted that this comparison was done using a mirror
of the kernel-series.yaml data to allow the times to be compared fairly.
Putting kernel-series.yaml from git.launchpad.net (the current default)
is a further 0.2s slower.

1: ./test7.py
	    Mean        Std.Dev.    Min         Median      Max
real        1.606       0.162       1.424       1.548       1.896
user        1.056       0.033       1.028       1.040       1.137
sys         0.026       0.004       0.020       0.024       0.036

Signed-off-by: Andy Whitcroft <apw@canonical.com>