cranky: update-dkms-versions -- use c prefix for embargoed updates
Harmonize update-dkms-versions to kernel packaging commit

"UBUNTU: [Packaging] convert to v7.4 autogen form"

Update the pattern used to determine which remote to use as a reference.
The c prefix now means the cycle is for embargoed updates which use a
private repo. The s prefix now means non-embargoes updates which use a
public repo.

Signed-off-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>