cranky/link-tb: Reduce the number of requests done
We are currently requesting all bugs for `start-sru-cycle` (500
at the time of writing) and we filter them after based on the
name including `<target-package>`.

But this generate a lot of useless requests. We can directly
request our `<target-package>` when requesting bugs to Launchpad.
This reduce the number of request by a factor ~10 when targeting
a derivative.

Later, when the `<series>/` prefix become mandatory, we will be able
to include it in the filter to reach a better accuracy.

Signed-off-by: Thibault Ferrante <thibault.ferrante@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>
Signed-off-by: Thibault Ferrante <thibault.ferrante@canonical.com>