adt-matrix: show only one retry link
When a package update is detected, show a retry link for only the new
version. The retry symbol prefixed with a + sign means the package was
updated and that the retry url will route to the new package. Duplicate
the tooltip annotation under the link so the user has more chances to
see what is going on.

Also remove the XXX todo commentary related to this feature.

Signed-off-by: Cory Todd <cory.todd@canonical.com>