cranky: crl/handle -- add cycle awareness to lookups
When instantiating HandleSet objects we use KernelSeries information
to determine the appropriate dependent packages in the set.  As this
may change over time it is important to instantiate a cycle appropriate
KernelSeries when making this determination.

Add an optional cycle parameter to lookup functions, this is used to
guide the selection of KernelSeries data.  When identifying handles via
a supplied path to a directory we additionally use the contained cycle
information in the tracking-bug to guild the default selection.  If both
are specified the caller supplied data takes precidence.

Signed-off-by: Andy Whitcroft <apw@canonical.com>