stable-tools: Add intial stable patchset scripts to stable-tools
Add git-stable-extag and stable-check-reverts into a top level directory
where we can put tools for processing stable patchsets and other
stable related tasks.

This would differ from the "stable" directory in that it would be for tools
not related directly to swm and cycle management.

Ported from Kamal's stable tools.

Signed-off-by: Luke Nowakowski-Krijger <luke.nowakowskikrijger@canonical.com>
Acked-by: kamalmostafa <kamal@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>