kernel-series: EOL Kinetic Kudu
Mark 22.10/Kinetic as EOL. Also change jammy:linux-hwe-5.19 to be a
primary kernel on its own and all former Kinetic backports to be
derivatives of hwe-5.19.
This will require some manual adjustments (rebase, version numbers) when
cranking the next time but is required to be able to maintain the Jammy
kernels beyond the EOL of Kinetic.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>