chroot-setup: add libfuse-dev as dependency for selftests
Compiling selftests on the tests builders which is done automatically
after each update on the repos, will fail with the following error:

fuse_mnt.c:17:10: fatal error: fuse.h: No such file or directory
    17 | #include <fuse.h>
       |          ^~~~~~~~

due to missing libfuse-dev.
Even though this was fixed by manually updating the existing chroots,
this change is intented to fix the issue for future releases.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>