cranky: checkout -- CycleRef generate should use per tree remotes
When we generate cycle-refs we need to fetch all outstanding tags in
each repository.  It is possible for the remote names to be different in
each package repository (however undesirable this is).  Switch to using
per tree rmeote names for this fetch.  Take this oppotunity to raise the
visibilty of these fetches as they are potentially very slow.

Signed-off-by: Andy Whitcroft <apw@canonical.com>