annotations-tools: drop kernelconfig
Drop kernelconfig from annotations-tools. This script is not portable
across multiple releases, so we don't want to sync it via `cranky fix`.

Remove it for now from the centralized repository.

In the future we may decide to fix this script to make it portable
across all the releases and then we can re-add it back, but for now it
is better to exclude it to avoid potential breakage during the cranky
process.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>