info/kernel-series: Set all swm.kernel-testing: true
The `swm.kernel-testing` flag is used to generate test requests with a
nice `kernel-testing--{soure name}--full--{flavour}` name instead of the
package source name. This can disambiguate package names in certain
cases and should be the perferred method. Update all the cases in
info/kernel-series.yaml where we've added kernels with this flag set to
false. We should no longer have kernels that are not built with these
virtual package names.

Signed-off-by: Francis Ginther <francis.ginther@canonical.com>