ktl/kernel_series: add support for cycles under KERNEL_SERIES_USE=local
Add support for picking cycle specific information from a locally places
kernel-versions repository under KERNEL_SERIES_USE=local.  It is assumed
that a static checkout of the kernel-versions repo is located in the
info directory.  We also support locally placed cycle files under
info/kernel-series.yaml@<cycle> for tight local testing.

We switch to producing a list of possible locations for the data in each
mode.  We then scan those ignoring remote 404 and local ENOENT errors
until we find the data or run out of options.

Signed-off-by: Andy Whitcroft <apw@canonical.com>