ktl/kernel-series: expose key_series_name on KernelSeriesUrl
We have (horribly) exposed a sort order key function for series called
key_series_name.  Normally this is used as a classmethod of
KernelSeries:

    for kss in sorted(ks.series, key=KernelSeries.key_series_name):

However there are also instances of callers doing:

    ks = KernelSeries()
    for kss in sorted(ks.series, key=ks.key_series_name):

Since the recent rework which introduced KernelSeriesCache and
KernelSeriesUrl the object returned when envoking KernelSeries() is now
a KernelSeriesUrl which does not have this class method.  Provide this
method in both places.

Signed-off-by: Andy Whitcroft <apw@canonical.com>