ancillary-bot: retry in the face of download/extraction failures
As we know we are downloading source packages which have been accepted
by the archive it is in theory impossible for those to be malformed.
Any download or extraction failures are more likley related to a failure
to fetch the correct contents.  In those cases mark the id for "retry"
which triggers the full removal and a reattempt on the next cycle.

Signed-off-by: Andy Whitcroft <apw@canonical.com>