swm: manager -- use kernel-sru-workflow status to indicate closed
We currently use the phase == Complete to indicate that a tracker is
closed.  This can be out of sync when we duplicate a tracker.  Switch to
using the top-level kernel-sru-workflow task as this is the real marker.
When this is "Fix Committed" or "Fix Released" the tracker is truly
closed.

Signed-off-by: Andy Whitcroft <apw@canonical.com>