gen-sru-announce: refactor SRU cycles table
Currently the SRU cycles table is created adding for each cycle a new
table with the cycle dates. This makes it hard to expand and align the
rows as the cycle tables are independent. Refactor how the table is
created, by expanding it adding rows instead of columns.

Also update the css styles in k.u.c index page to match the new table
layout.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>