docs: s/linux-DERIVATIVE-restricted-modules/linux-lrm/
After running a `cranky checkout lunar:linux-kvm`, the
`linux-kvm-restricted-modules/` directory is not created but instead
the `linux-lrm/` folder name is used. After asking to the team,
this seems to be the new name in use for restricted modules.
Also, the `-DERIVATIVE` suffix is gone.

Reflect these name changes into the docs.

Note: this change makes the docs inconsistent, as linux-oracle
still seems to use the `linux-restricted-modules` name.

TODO: use something else than linux-oracle for the examples.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>
Signed-off-by: Agathe Porte <agathe.porte@canonical.com>