create-kernel-tasks: Work around spin 1 Launchpad bug
When looking up workflow trackers to check for exising ones there is a
difference between spin #1 and later ones. The reason why is somewhat
lost in history. While for later spins the library is called with a set
of all cycle tags, we pass an empty list for the initial spin.

This gets translated by the library into a search for all workflow tasks
which have a "kernel-release-tracking-bug-live" tag. Right now this
seems to return about 4000-5000 entries when done via the web page. The
API call however runs into a timeout and returns a Launchpad Oops.

Right now the simplest solution is to modify create-kernel-tasks to pass
in the cycle tag for spin #1 as well. Which also speeds up the runtime a
lot.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>