cranky: crl/handle -- include cycle name in lookup failures
When cranky is asked to work on a tree set that depends on changes not
represented in the effective kernel-series cycle an incomplete error
message is shown. Include the cycle name in the error as a hint to the
user that they should either fix the kernel-series@cycle or use the
correct cycle.

Signed-off-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>