cranky: open: do not modify state when combining --dry-run and --reuse-abi
cranky open --reuse-abi modifies $DEBIAN/abi/version and adds it to the
index irrespectively whether --dry-run is passed or not. This will also
cause future invocations of cranky open to refuse to run since $DEBIAN
will not be clean.

Change the script to respect --dry-run in all cases and leave the tree
unaltered without having to reset the state of the tree.

Signed-off-by: Manuel Diewald <manuel.diewald@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Manuel Diewald <manuel.diewald@canonical.com>