stable-tools: stable-check-reverts: enable remote upstream branches
The script assumes to be run inside an upstream stable repository. To
allow for a different setup, e.g. where upstream stable is only
available as a remote to a local repository, try to find the correct
upstream branch in a more sophisticated way. In case there is more than
one branch that fits the pattern, select the first one of the list.

This approach allows for the traditional setup but also enables a
different repository layout.

Signed-off-by: Manuel Diewald <manuel.diewald@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Manuel Diewald <manuel.diewald@canonical.com>