cranky: review-master-changes -- fix 2nd order derivatives
In some cases such as respins, it was possible to not be able to find a
starting point for calculating a changelist. This was addressed in

29ed91b7 ("cranky: review-master-changes: Fix the case when the base in the changelog is not in the tree")

This approach was found to not work with derivates of derivaties. Let's
use the original method and then attempt the more expanded method as a
fallback.

Signed-off-by: Cory Todd <cory.todd@canonical.com>
Acked-by: John Cabaj <john.cabaj@canonical.com>
Signed-off-by: Cory Todd <cory.todd@canonical.com>