stable/sru-cycle-kernel-versions: Ignore security cycle
This is called in the `gen-sru-announce` script used only for SRU cycle
annoucements at the moment to get the latest main kernels versions:
    `sru-cycle-kernel-versions --no-devel --no-esm`

It checks https://kernel.ubuntu.com/sru/kernel-sru-workflow.json for
the current cycle(s) kernels that are non esm and non development.
If there is a security cycle overlapping, 2 kernels are found and the
security version may be the one returned in case its tracker is newer.
This fixes it by checking if the cycle number starts with 's' and ignoring that
version in that case.

Note: This is a quick fix and if we need to make announcements for
security we will have to extend this script to accomodate that.

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>