swm: snap recipies v2 -- add interfaces to find the new snap recipies
Add new interfaces to identify and load data from the new V2 format snap
recipies.  These are recipies which expect SWM to take the lead when
triggeing builds.  We do this so that SWM can keep a handle on the
request-builds and better identify the resulting builds and revisions in
the store.

Signed-off-by: Andy Whitcroft <apw@canonical.com>