stable/24hrs-of-bugs: Remove cb() class method and 'data' class variable
Currently, a class method is used as a callback (why?) to accumulate
results of multiple searches and store them in the 'data' class variable.
They don't need to be class members since they don't access any real
class resources. It is really convoluted, an abuse of the class concept
and makes no sense.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>