swm: swm-worker -- ack away failed requests
When we complete a request ack it way regardless of the result.  All
requests will be resubmitted if they are not actioned and as a result
if we leave failed ones on the queue we will simply block the queue.
Rely on the oops reporting to let us notice trackers not making
progress.

Signed-off-by: Andy Whitcroft <apw@canonical.com>