swm: task -- abort on excessive task updates
We have gotten into a situation where we do excessive task updates.
When this occurs we will spam the activity log of the associated bug to
the point we cannot use the bug any longer and database surgery can be
needed.  Cap out task updates at 100, which more than 3 per live task.

Signed-off-by: Andy Whitcroft <apw@canonical.com>