stable-tools: Add script to check for existing critical fixes
This tool is meant to be run when preparing an upstream stable patchset.
It will check for potentially critical fixes in future patchsets on the
same branch, that fix issues in patchset you are currently working on.
What is considered critical is based on a set of regular expressions
that might produce some false positives and miss actually critical
fixes.

The script will present critical fixes it detected in a nice way with
some additional information. The user may then decide whether it is
reasonable to apply the fix ahead of time.

Ideally we would be able to avoid applying patches that severely break
the kernel even though a fix is readily available.

Signed-off-by: Manuel Diewald <manuel.diewald@canonical.com>