cranky: pull-sources: new -l, --latest option to ignore .getabis
While cranking regular kernels, it is cumbersome to create .getabis
files just for pulling all sources. The pull-source command supports
automatic pull without the .getabis files if no version is passed.

Introduce a new -l, --latest option to make pull-sources not use
.getabis but do not enable this option by default. This will facilitate
the cranking process for regular kernels, because the current process if
one does not want to create .getabis files is to manually call
pull-source for each package.

Signed-off-by: Agathe Porte <agathe.porte@canonical.com>