cranky: chroot: install libtraceevent-dev for mantic, and jammy
The package libtraceevent-dev is needed for the m:linux-intel-opt
kernel to enable "perf kvm stat" this package is required for mantic.

This is required for intel TDX features.

The package has also been added to jammy because of the m:intel-opt
kernel has been backported to jammy, so it is needed in jammy to build
the backports.

Signed-off-by: Philip Cox <philip.cox@canonical.com>