cranky: Add new update-dkms-versions subcommand
The debian/dkms-versions file is in the wrong location, it should live
in debian.<variant>. Rather than fixing the update-dkms-versions
script in cranky/master and fix it into all trees, pull it out and
make it a cranky subcommand.

Note that this subcommand can only be used on trees that don't have the
in-tree update-dkms-version script anymore. This in-tree script needs to
be manually and explicitly removed, to ensure that we can move one tree
at a time. But also note that once the in-tree script is removed from a
main/master kernel, all the derivatives inherit the process change.

Jira: https://warthogs.atlassian.net/browse/KERNTT-709
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>