cranky: Add new update-dependent-version subcommand
All dependent packages that need to be manually cranked have an in-tree
update-version script which is called from the update-dependent
subcommand. For the meta and lrm packages, this script pulls over the
dkms-versions file from the main tree. Given that that file moved to
a new location debian.<variant>, the update-version script obviously
needs to be modified. However, every dependent package has its own
update-version script and they have all diverged slightly over time and
are different now (sigh). Rather than cranky fix all of that mess,
create a new subcommand update-dependent-version that can be used for
all dependent packages.

Due to the differences between package types and the divergence over
time, this has potential for regression. So the produced output needs
to be carefully reviewed. Also note that this subcommand can only
be used after the in-tree update-version script has been removed. Again,
this ensures a controlled switch over of all the trees.

Jira: https://warthogs.atlassian.net/browse/KERNTT-709
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>