ktl/sru_cycle: Fix default local URL and add YAML dump support
The local URL usage got broken when we moved the sru-cycle.yaml file.
This now is part of a different repository. For simplicity require that
the other repository is located in the same root as kteam-tools and is
named 'kernel-versions'.

Add a method to convert SruCycle and SruCycleSpinEntry into strings.
When used on SruCycle this will generate the informational header and
then the YAML definition of every cycle/spin entry from latest to
oldest.

To have the string conversion produce consistent output a sort key for
cycle/spin names is added. This will first order ascending by date. The
various cycle types get ordered by SRU, security, and devel. And within
the same cycle the final order is spin number. A cycle entry without
spin number is treated as if it had spin number 0.

As a fix-up "cutoff-date" had to be added to the SruCycleSpinEntry
parser to prevent it being lost.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>