ktl/sru_cycle: Allow adding a cycle
The way this is done is a bit weird but it seemed to me the simplest
form of re-using present code. Most the logic about an individual cycle
is done in SruCycleSpinEntry but the collection contained in SruCycle
consists of a dictionary. So I go through creating an entry object
first and then merge it via YAML -> dictionary conversion.

Added two testcases, one for checking that no duplicates get added and
the other to see a new cycle gets added normally.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>