cranky: open/close: Use cranky rather than calling make targets
The open/close helpers used to live in the trees and were directly
calling make targets that run in a chroot session. A helper-wrapper
script was defining the chroot session to use. The helper-wrapper is
gone so that no longer works correctly. Use cranky instead (which is
the preferred way anyways).

https://warthogs.atlassian.net/browse/KERNTT-646
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>