ktl/sru_cycle: Add write and delete_cycle
Add a function to SruCycle which allows writing the current dataset back
to a "local:" URL (file).

Also add a delete_cycle() function as the counter-part of add_cycle().

In addition document lookup_cycle() and lookup_spin() and try to have
sane mappings for special cases where looking for a spin might return a
cycle (this already was done before) or looking for a cycle might return
a spin.

Drop some unused code and also harden loading from an URL. Failure now
returns an empty dataset. Maybe that would have to be limited to the
case of a file:/// URL pointing to a non-existing file.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>