ktl/new_release: add class to bump package version
This is a python implementation of the startnewrelease algorithm. It does not
need the suffix, figuring it out from the version itself. It does need the
parent version for the cases where it has changed (a rebase has been done, for
example).

v2:
Raise ValueError if an unexpected version format is found.
Document the 3 different scenarios.
Add test for invalid cases, including the ones still not considered invalid.

v3 [juergh]:
Add new files to pre-commit whitelist and accept 'black' changes

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>