cranky/cmds.d/rebase*: drop automatically updates to auto-sync files
cranky-fix keeps a number of files in sync with master copies.  Evaluate
the local delta and drop any updates to those files before rebasing to
avoid collisions with those files; this is safe because we know the post
rebase cranky-fix will ensure they are repaired.

Also if the rebase renders a commit empty request it is automatically
dropped.

Signed-off-by: Andy Whitcroft <apw@canonical.com>