cranky: fix: Remove update-version and update-dkms-versions from dependent trees
If the main tree doesn't have the update-dkms-version script anymore, we
need to remove the in-tree update scripts from the dependents as well.
They can't handle the new location of the dkms-versions file and the
new cranky subcommands need to be used instead.

Jira: https://warthogs.atlassian.net/browse/KERNTT-709
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>