cranky: chroot: add which handling
Users can specify chroot session names (appended to the normal chroot
prefixes) using the --which option, like the old <sources|configs>. If
left out, it will create a default session.

The kernel series can be specified with: only the series, the handle,
or a path. If the series is omitted, it tries to infer the series from
the current directory.

Some help messages have been updated accordingly.

https://warthogs.atlassian.net/browse/KERNTT-724
Signed-off-by: Noah Wager <noah.wager@canonical.com>
[juergh: Limited the commit to adding --which commandline argument,
 trimmed the commit message.]
Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>