stable/create-cycle-sprint: Use sru_cycle library to update cycle info
This replaces the hard coded update of the sru-cycle.yaml file by usage
of the updated ktl/sru_cycle.py library. Which implicitly fixes the
issue of the file being relocated.

This adds a new requirement for having 'kernel-versions' checked out
with that name in the same top-level directory as 'kteam-tools' is.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>