stable/create-cycle-spint: Add notes-link to cycle info
With the new "Cycle README" issues being created, take a reference to
those when updating the cycle info.
To get the reference to the Jira card, the code to create the cards and
issues had to move to be done first. This however delays the check on a
duplicated cycle name and by then an existing board might be "polluted"
by duplicated issues. So an explicit test got added before starting to
make the board and issues.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>