adt-matrix: Include main linux package version in triggers for retry
For dkms modules that do not have any specific tests, the matrix will run
`dkms-autopkgtest` (part of dkms source) which consists of building the
module only. The test will run only if the `trigger` parameter contains
the linux main package version because it expects the following version
format:
linux-<derivative>/<kernel_version>-<abi_version>.<upload_version>

For retry, the `trigger` param consists of only meta version which has the
following format:
linux-<derivative>/<kernel_version>.<abi_version>.<upload_version>
which contains a dot instead of a '-' and prevents the test from
running and it marks it as 'Passed'.

Fix this by including the main version for retries too.
Note: When the test is triggered for the first time, the main version is
included, that's why the first run is always correct. The problem appears
only for retries. This broken logic prevents dkms builds failures from
being properly detected if the test is rerun (which usually happens because
people try the test 1-2 times before confirming if there's an issue)

Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>