ktl/new_release: handle case when parent version is lower
focal:linux-riscv-5.15 has jammy:linux as parent. It has an extra backport
prefix, but it also bumps its own ABI, which is higher than the parent. So it
is a mix of a derivative and a backport.

Currently, _bump_main handles that case by reusing the parent version,
which leads to a lower bumped version.

Fix that by ignoring the parent version when it is lower than the main
kernel version and resetting the last digit from EXTRA part and
incrementing both ABI and UPLOAD numbers.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>