cves-jira-next-cycle: attach a cycle field to certain CVE cards
Certain CVE JIRA issues will have a 'next' label attached to it, meaning
they are targeted to the next kernel release cycle. But that doesn't tell
which cycle it is, which might be confusing when looking for the CVEs that
need to go into a cycle.

cves-jira-next-cycle will look for a 'Cycle' meta field, a comment starting
with 'Cycle: ' and extract the cycle from it and list all CVEs belonging to
a given cycle that have the 'next' label.

It will also require that a cycle is given in the command line and any CVE
issue with the 'next' label that doesn't have a cycle meta field will have
that meta field added with that given cycle.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Cengiz Can <cengiz.can@canonical.com>
Signed-off-by: Cengiz Can <cengiz.can@canonical.com>